Make first extraction of member page
